it's vore station

+++ code/modules/power/tesla/energy_ball.dm
@@ -23,23 +23,26 @@
 	var/energy_to_raise = 32
 	var/energy_to_lower = -20
 
-/obj/singularity/energy_ball/Initialize(mapload, starting_energy = 50, is_miniball = FALSE)
+/obj/singularity/energy_ball/New(loc, starting_energy = 50, is_miniball = FALSE)
+	..()
 	miniball = is_miniball
+
+/obj/singularity/energy_ball/Initialize()
 	. = ..()
-	if(!is_miniball)
+	if(!miniball)
 		set_light(10, 7, "#EEEEFF")
 
 /obj/singularity/energy_ball/ex_act(severity, target)
 	return
 
 /obj/singularity/energy_ball/Destroy()
-	if(orbiting && istype(orbiting.parent, /obj/singularity/energy_ball))
-		var/obj/singularity/energy_ball/EB = orbiting.parent
+	if(orbiting && istype(orbiting.orbiting, /obj/singularity/energy_ball))
+		var/obj/singularity/energy_ball/EB = orbiting.orbiting
 		EB.orbiting_balls -= src
 
 	for(var/ball in orbiting_balls)
 		var/obj/singularity/energy_ball/EB = ball
-		QDEL_NULL(EB)
+		qdel(EB)
 
 	. = ..()
 
@@ -49,55 +52,59 @@
 	..()
 
 
-/obj/singularity/energy_ball/process()
+/obj/singularity/energy_ball/process(var/wait = 20)
+	set waitfor = FALSE
 	if(!orbiting)
-		handle_energy()
+		if (handle_energy())
+			return
 
-		move_the_basket_ball(4 + orbiting_balls.len * 1.5)
+		move_the_basket_ball(max(wait - 5, 4 + orbiting_balls.len * 1.5))
 
-		playsound(src.loc, 'sound/magic/lightningbolt.ogg', 100, TRUE, extrarange = 30)
+		playsound(src.loc, 'sound/effects/lightningbolt.ogg', 100, 1, extrarange = 30)
 
-		pixel_x = 0
-		pixel_y = 0
+		set_dir(tesla_zap(src, 7, TESLA_DEFAULT_POWER, TRUE))
 
-		tesla_zap(src, 7, TESLA_DEFAULT_POWER, TRUE)
-
-		pixel_x = -32
-		pixel_y = -32
 		for (var/ball in orbiting_balls)
 			var/range = rand(1, CLAMP(orbiting_balls.len, 3, 7))
-			tesla_zap(ball, range, TESLA_MINI_POWER/7*range)
+			tesla_zap(ball, range, TESLA_MINI_POWER/7*range, TRUE)
 	else
 		energy = 0 // ensure we dont have miniballs of miniballs
 
 /obj/singularity/energy_ball/examine(mob/user)
-	. = ..()
+	..()
 	if(orbiting_balls.len)
-		. += "There are [orbiting_balls.len] mini-balls orbiting it."
+		to_chat(user, "The amount of orbiting mini-balls is [orbiting_balls.len].")
 
 
 /obj/singularity/energy_ball/proc/move_the_basket_ball(var/move_amount)
 	//we face the last thing we zapped, so this lets us favor that direction a bit
-	var/move_bias = pick(GLOB.alldirs)
+	var/move_bias = dir
 	for(var/i in 0 to move_amount)
-		var/move_dir = pick(GLOB.alldirs + move_bias) //ensures large-ball teslas don't just sit around
+		var/move_dir = pick(global.alldirs + move_bias) //ensures large-ball teslas don't just sit around
 		if(target && prob(10))
 			move_dir = get_dir(src,target)
 		var/turf/T = get_step(src, move_dir)
 		if(can_move(T))
 			forceMove(T)
-			setDir(move_dir)
+			set_dir(move_dir)
 			for(var/mob/living/carbon/C in loc)
 				dust_mobs(C)
-
+			sleep(1) // So movement is smooth
 
 /obj/singularity/energy_ball/proc/handle_energy()
+	if (energy <= 0)
+		log_game("TESLA([x],[y],[z]) Collapsed entirely.")
+		investigate_log("collapsed.", I_SINGULO)
+		qdel(src)
+		return TRUE
+
 	if(energy >= energy_to_raise)
 		energy_to_lower = energy_to_raise - 20
 		energy_to_raise = energy_to_raise * 1.25
 
-		playsound(src.loc, 'sound/magic/lightning_chargeup.ogg', 100, TRUE, extrarange = 30)
-		addtimer(CALLBACK(src, .proc/new_mini_ball), 100)
+		playsound(src.loc, 'sound/effects/lightning_chargeup.ogg', 100, 1, extrarange = 30)
+		//addtimer(CALLBACK(src, .proc/new_mini_ball), 100)
+		spawn(100) new_mini_ball()
 
 	else if(energy < energy_to_lower && orbiting_balls.len)
 		energy_to_raise = energy_to_raise / 1.25
@@ -129,36 +136,27 @@
 /obj/singularity/energy_ball/Bumped(atom/movable/AM)
 	dust_mobs(AM)
 
-/obj/singularity/energy_ball/attack_tk(mob/user)
-	if(iscarbon(user))
-		var/mob/living/carbon/C = user
-		to_chat(C, "<span class='userdanger'>That was a shockingly dumb idea.</span>")
-		var/obj/item/organ/brain/rip_u = locate(/obj/item/organ/brain) in C.internal_organs
-		C.ghostize(0)
-		qdel(rip_u)
-		C.death()
-
 /obj/singularity/energy_ball/orbit(obj/singularity/energy_ball/target)
 	if (istype(target))
 		target.orbiting_balls += src
-		GLOB.poi_list -= src
+		//TODO-LESH-DEL global.poi_list -= src
 		target.dissipate_strength = target.orbiting_balls.len
 
 	. = ..()
 /obj/singularity/energy_ball/stop_orbit()
-	if (orbiting && istype(orbiting.parent, /obj/singularity/energy_ball))
-		var/obj/singularity/energy_ball/orbitingball = orbiting.parent
+	if (orbiting && istype(orbiting.orbiting, /obj/singularity/energy_ball))
+		var/obj/singularity/energy_ball/orbitingball = orbiting.orbiting
 		orbitingball.orbiting_balls -= src
 		orbitingball.dissipate_strength = orbitingball.orbiting_balls.len
-	. = ..()
-	if (!QDELETED(src))
+	..()
+	if (!loc && !QDELETED(src))
 		qdel(src)
 
 
 /obj/singularity/energy_ball/proc/dust_mobs(atom/A)
 	if(isliving(A))
 		var/mob/living/L = A
-		if(L.incorporeal_move || L.status_flags & GODMODE)
+		if(L.incorporeal_move)
 			return
 	if(!iscarbon(A))
 		return
@@ -166,9 +164,10 @@
 		if(GR.anchored)
 			return
 	var/mob/living/carbon/C = A
-	C.dust()
+	// C.dust() - Changing to do fatal elecrocution instead
+	C.electrocute_act(500, src, def_zone = BP_TORSO)
 
-/proc/tesla_zap(atom/source, zap_range = 3, power, tesla_flags = TESLA_DEFAULT_FLAGS, list/shocked_targets)
+/proc/tesla_zap(atom/source, zap_range = 3, power, explosive = FALSE, stun_mobs = TRUE)
 	. = source.dir
 	if(power < 1000)
 		return
@@ -182,10 +181,11 @@
 	var/obj/structure/closest_structure
 	var/obj/structure/blob/closest_blob
 	var/static/things_to_shock = typecacheof(list(/obj/machinery, /mob/living, /obj/structure))
-	var/static/blacklisted_tesla_types = typecacheof(list(/obj/machinery/atmospherics,
+	var/static/blacklisted_tesla_types = typecacheof(list(
+										/obj/machinery/atmospherics,
 										/obj/machinery/power/emitter,
-										/obj/machinery/field/generator,
-										/mob/living/simple_animal,
+										/obj/machinery/field_generator,
+										/obj/machinery/door/blast,
 										/obj/machinery/particle_accelerator/control_box,
 										/obj/structure/particle_accelerator/fuel_chamber,
 										/obj/structure/particle_accelerator/particle_emitter/center,
@@ -193,26 +193,19 @@
 										/obj/structure/particle_accelerator/particle_emitter/right,
 										/obj/structure/particle_accelerator/power_box,
 										/obj/structure/particle_accelerator/end_cap,
-										/obj/machinery/field/containment,
+										/obj/machinery/containment_field,
 										/obj/structure/disposalpipe,
-										/obj/structure/disposaloutlet,
-										/obj/machinery/disposal/deliveryChute,
-										/obj/machinery/camera,
 										/obj/structure/sign,
 										/obj/machinery/gateway,
 										/obj/structure/lattice,
 										/obj/structure/grille,
-										/obj/machinery/the_singularitygen/tesla,
-										/obj/structure/frame/machine))
+										/obj/machinery/the_singularitygen/tesla))
 
 	for(var/A in typecache_filter_multi_list_exclusion(oview(source, zap_range+2), things_to_shock, blacklisted_tesla_types))
-		if(!(tesla_flags & TESLA_ALLOW_DUPLICATES) && LAZYACCESS(shocked_targets, A))
-			continue
-
 		if(istype(A, /obj/machinery/power/tesla_coil))
 			var/dist = get_dist(source, A)
 			var/obj/machinery/power/tesla_coil/C = A
-			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !(C.obj_flags & BEING_SHOCKED))
+			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !C.being_shocked)
 				closest_dist = dist
 
 				//we use both of these to save on istype and typecasting overhead later on
@@ -225,7 +218,8 @@
 			continue //no need checking these other things
 
 		else if(istype(A, /obj/machinery/power/grounding_rod))
-			var/dist = get_dist(source, A)-2
+			var/obj/machinery/power/grounding_rod/G = A
+			var/dist = get_dist(source, A) - (G.anchored ? 2 : 0)
 			if(dist <= zap_range && (dist < closest_dist || !closest_grounding_rod))
 				closest_grounding_rod = A
 				closest_atom = A
@@ -237,7 +231,7 @@
 		else if(isliving(A))
 			var/dist = get_dist(source, A)
 			var/mob/living/L = A
-			if(dist <= zap_range && (dist < closest_dist || !closest_mob) && L.stat != DEAD && !(L.flags_1 & TESLA_IGNORE_1))
+			if(dist <= zap_range && (dist < closest_dist || !closest_mob) && L.stat != DEAD && !(L.status_flags & GODMODE))
 				closest_mob = L
 				closest_atom = A
 				closest_dist = dist
@@ -245,10 +239,10 @@
 		else if(closest_mob)
 			continue
 
-		else if(ismachinery(A))
+		else if(istype(A, /obj/machinery))
 			var/obj/machinery/M = A
 			var/dist = get_dist(source, A)
-			if(dist <= zap_range && (dist < closest_dist || !closest_machine) && !(M.obj_flags & BEING_SHOCKED))
+			if(dist <= zap_range && (dist < closest_dist || !closest_machine) && !M.being_shocked)
 				closest_machine = M
 				closest_atom = A
 				closest_dist = dist
@@ -259,7 +253,7 @@
 		else if(istype(A, /obj/structure/blob))
 			var/obj/structure/blob/B = A
 			var/dist = get_dist(source, A)
-			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !(B.obj_flags & BEING_SHOCKED))
+			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !B.being_shocked)
 				closest_blob = B
 				closest_atom = A
 				closest_dist = dist
@@ -267,50 +261,60 @@
 		else if(closest_blob)
 			continue
 
-		else if(isstructure(A))
+		else if(istype(A, /obj/structure))
 			var/obj/structure/S = A
 			var/dist = get_dist(source, A)
-			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !(S.obj_flags & BEING_SHOCKED))
+			if(dist <= zap_range && (dist < closest_dist || !closest_tesla_coil) && !S.being_shocked)
 				closest_structure = S
 				closest_atom = A
 				closest_dist = dist
-				
-		else if(closest_structure)
-			continue
 
 	//Alright, we've done our loop, now lets see if was anything interesting in range
 	if(closest_atom)
 		//common stuff
-		source.Beam(closest_atom, icon_state="lightning[rand(1,12)]", time=5, maxdistance = INFINITY)
-		if(!(tesla_flags & TESLA_ALLOW_DUPLICATES))
-			LAZYSET(shocked_targets, closest_atom, TRUE)
+		var/atom/srcLoc = get_turf(source) // VOREStation Edit - Makes beams look nicer
+		srcLoc.Beam(closest_atom, icon_state="lightning[rand(1,12)]", time=5, maxdistance = INFINITY)  // VOREStation Edit - Makes beams look nicer
 		var/zapdir = get_dir(source, closest_atom)
 		if(zapdir)
 			. = zapdir
 
+	var/drain_energy = FALSE // VOREStation Edit - Safety First! Drain Tesla fast when its loose
+
 	//per type stuff:
-	if(!QDELETED(closest_tesla_coil))
-		closest_tesla_coil.tesla_act(power, tesla_flags, shocked_targets)
+	if(closest_tesla_coil)
+		closest_tesla_coil.tesla_act(power, explosive, stun_mobs)
 
-	else if(!QDELETED(closest_grounding_rod))
-		closest_grounding_rod.tesla_act(power, tesla_flags, shocked_targets)
+	else if(closest_grounding_rod)
+		closest_grounding_rod.tesla_act(power, explosive, stun_mobs)
 
-	else if(!QDELETED(closest_mob))
-		var/shock_damage = (tesla_flags & TESLA_MOB_DAMAGE)? (min(round(power/600), 90) + rand(-5, 5)) : 0
-		closest_mob.electrocute_act(shock_damage, source, 1, SHOCK_TESLA | ((tesla_flags & TESLA_MOB_STUN) ? NONE : SHOCK_NOSTUN))
+	else if(closest_mob)
+		var/shock_damage = CLAMP(round(power/400), 10, 90) + rand(-5, 5)
+		closest_mob.electrocute_act(shock_damage, source, 1 - closest_mob.get_shock_protection(), ran_zone())
+		log_game("TESLA([source.x],[source.y],[source.z]) Shocked [key_name(closest_mob)] for [shock_damage]dmg.")
+		message_admins("Tesla zapped [key_name_admin(closest_mob)]!")
 		if(issilicon(closest_mob))
 			var/mob/living/silicon/S = closest_mob
-			if((tesla_flags & TESLA_MOB_STUN) && (tesla_flags & TESLA_MOB_DAMAGE))
-				S.emp_act(EMP_LIGHT)
-			tesla_zap(S, 7, power / 1.5, tesla_flags, shocked_targets) // metallic folks bounce it further
+			if(stun_mobs)
+				S.emp_act(3 /*EMP_LIGHT*/)
+			tesla_zap(closest_mob, 7, power / 1.5, explosive, stun_mobs) // metallic folks bounce it further
 		else
-			tesla_zap(closest_mob, 5, power / 1.5, tesla_flags, shocked_targets)
+			tesla_zap(closest_mob, 5, power / 1.5, explosive, stun_mobs)
 
-	else if(!QDELETED(closest_machine))
-		closest_machine.tesla_act(power, tesla_flags, shocked_targets)
+	else if(closest_machine)
+		drain_energy = TRUE // VOREStation Edit - Safety First! Drain Tesla fast when its loose
+		closest_machine.tesla_act(power, explosive, stun_mobs)
 
-	else if(!QDELETED(closest_blob))
-		closest_blob.tesla_act(power, tesla_flags, shocked_targets)
+	else if(closest_blob)
+		drain_energy = TRUE // VOREStation Edit - Safety First! Drain Tesla fast when its loose
+		closest_blob.tesla_act(power, explosive, stun_mobs)
 
-	else if(!QDELETED(closest_structure))
-		closest_structure.tesla_act(power, tesla_flags, shocked_targets)
+	else if(closest_structure)
+		drain_energy = TRUE // VOREStation Edit - Safety First! Drain Tesla fast when its loose
+		closest_structure.tesla_act(power, explosive, stun_mobs)
+
+	// VOREStation Edit Start - Safety First! Drain Tesla fast when its loose
+	if(drain_energy && istype(source, /obj/singularity/energy_ball))
+		var/obj/singularity/energy_ball/EB = source
+		if (EB.energy > 0)
+			EB.energy -= min(EB.energy, max(10, round(EB.energy * 0.05)))
+	// VOREStation Edit End
